moved a few classes to other packages
