--- src/main/scala/com/truelaurel/algorithm/mcts/perf/MctsDebugPerf.scala
@@ -1,10 +1,0 @@
-package com.truelaurel.algorithm.mcts.perf
-
-import com.truelaurel.algorithm.mcts.perf.DummyMcts._
-
-object MctsDebugPerf {
-  def main(args: Array[String]): Unit = {
-    dummyRoot.steps(1000000)
-  }
-
-}
