+++ src/main/scala/com/truelaurel/algorithm/perf/MctsDebugPerf.scala
@@ -1,0 +1,8 @@
+package com.truelaurel.algorithm.perf
+
+object MctsDebugPerf {
+  def main(args: Array[String]): Unit = {
+    DummyMcts.dummyRoot.steps(1000000)
+  }
+
+}
