+++ src/main/scala/com/truelaurel/codingame/tool/benchmark/HelloWorldBenchmark.scala
@@ -2,6 +2,7 @@ package com.truelaurel.codingame.tool.benchmark
 
 import org.openjdk.jmh.annotations.Benchmark
 
+// This class is there to check during CI that jmh benchmarks are not broken in SBT.
 class HelloWorldBenchmark {
 
   @Benchmark
