+++ src/main/scala/com/truelaurel/math/geometry/Vectorl.scala
@@ -1,4 +1,4 @@
-package com.truelaurel.math.vectorial
+package com.truelaurel.math.geometry
 
 import com.truelaurel.math.Mathl
